Remove root window tracking
This is really only applicable to X11, and passing it around though all
the function isn't really needed anyway. The only place it was used was
in the watch/unwatch functions, but we can just pass NULL to say that we
want all events instead of events for a particular window.