Insert windows into WM immediately
This simplifies later processing because we don't have to figure out if
we need to insert or remove windows when they are mapped or unmapped. We
can just insert and remove them when the are created and destroyed then
send ST_SHOW/ST_HIDE events the rest of the time.